arch-logos: Show all the logos of arch logo competition on one page.
This new page is just a small static page that shows all the various
logo submissions in a easier to browse fashion. We add some new node
packages to help with the job.

One good thing about this build system is I don't have to care how
slow the build is as long as the final result is fast so I'm happy
with how this page turned out.

Architecturally, now each page can have a server.js file that will
return some data to be passed to the template before it is rendered.
This means we don't have to rely on AJAX on the client side OR
hardcode in our code the images we want to fetch from the remote
server here (and we also bypass CORS, oops).